Tidied PR up a little bit.
